+++ Package.swift
@@ -6,6 +6,6 @@ let package = Package(
     name: "EasyMapping",
     products: [.library(name: "EasyMapping", targets: ["EasyMapping"])],
     targets: [
-        .target(name: "EasyMapping", exclude: ["Sources/EasyMapping+XCTestCase"])
+        .target(name: "EasyMapping", publicHeadersPath: "")
     ]
 )